JSON*.swift -> Swifter*.swift, remove wrapper
