+++ Sources/SwifterFavorites.swift
@@ -1,5 +1,5 @@
 //
-//  JSONFavorites.swift
+//  SwifterFavorites.swift
 //  Swifter
 //
 //  Copyright (c) 2014 Matt Donnelly.
