+++ Sources/SwifterFollowers.swift
@@ -1,5 +1,5 @@
 //
-//  JSONFollowers.swift
+//  SwifterFollowers.swift
 //  Swifter
 //
 //  Copyright (c) 2014 Matt Donnelly.
