+++ Sources/SwifterHelp.swift
@@ -1,5 +1,5 @@
 //
-//  JSONHelp.swift
+//  SwifterHelp.swift
 //  Swifter
 //
 //  Copyright (c) 2014 Matt Donnelly.
