+++ Sources/SwifterLists.swift
@@ -1,5 +1,5 @@
 //
-//  JSONLists.swift
+//  SwifterLists.swift
 //  Swifter
 //
 //  Copyright (c) 2014 Matt Donnelly.
