+++ Sources/SwifterMessages.swift
@@ -1,5 +1,5 @@
 //
-//  JSONMessages.swift
+//  SwifterMessages.swift
 //  Swifter
 //
 //  Copyright (c) 2014 Matt Donnelly.
