+++ Sources/SwifterSavedSearches.swift
@@ -1,5 +1,5 @@
 //
-//  JSONSavedSearches.swift
+//  SwifterSavedSearches.swift
 //  Swifter
 //
 //  Copyright (c) 2014 Matt Donnelly.
