+++ Sources/SwifterSpam.swift
@@ -1,5 +1,5 @@
 //
-//  JSONSpam.swift
+//  SwifterSpam.swift
 //  Swifter
 //
 //  Copyright (c) 2014 Matt Donnelly.
