+++ Sources/SwifterStreaming.swift
@@ -1,5 +1,5 @@
 //
-//  JSONStreaming.swift
+//  SwifterStreaming.swift
 //  Swifter
 //
 //  Copyright (c) 2014 Matt Donnelly.
