+++ Sources/SwifterTimelines.swift
@@ -1,5 +1,5 @@
 //
-//  JSONTimelines.swift
+//  SwifterTimelines.swift
 //  Swifter
 //
 //  Copyright (c) 2014 Matt Donnelly.
