--- Tests/SwifterTests/SwifterTests.swift
@@ -1,41 +1,0 @@
-//
-//  SwifterTests.swift
-//  SwifterTests
-//
-//  Created by Lewis Smith on 22/02/2016.
-//  Copyright © 2016 Matt Donnelly. All rights reserved.
-//
-
-import XCTest
-
-class SwifterTests: XCTestCase {
-    
-    override func setUp() {
-        super.setUp()
-        // Put setup code here. This method is called before the invocation of each test method in the class.
-    }
-    
-    override func tearDown() {
-        // Put teardown code here. This method is called after the invocation of each test method in the class.
-        super.tearDown()
-    }
-    
-    func testExample() {
-        // This is an example of a functional test case.
-        // Use XCTAssert and related functions to verify your tests produce the correct results.
-    }
-    
-    func testPerformanceExample() {
-        // This is an example of a performance test case.
-        self.measure {
-            // Put the code you want to measure the time of here.
-        }
-    }
-
-    #if os(Linux)
-    static var allTests = [
-        ("testExample", testExample),
-        ("testPerformanceExample", testPerformanceExample)
-    ]
-    #endif
-}
